% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{dRatioOfQsprays}
\alias{dRatioOfQsprays}
\title{Partial differentiation}
\usage{
dRatioOfQsprays(roq, orders)
}
\arguments{
\item{orders}{integer vector, the orders of the differentiation}

\item{qspray}{object of class \code{ratioOfQsprays}}
}
\value{
A \code{ratioOfQsprays} object.
}
\description{
Partial differentiation of a ratioOfQsprays polynomial.
}
\examples{
library(ratioOfQsprays)
x <- qlone(1)
y <- qlone(2)
roq <- (x + 2*y  + 3*x*y) / (x + 1)
dRatioOfQsprays(qspray, c(1, 1))
derivRatioOfQsprays(derivRatioOfQsprays(roq, 1), 2)
}
