% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{getDenominator}
\alias{getDenominator}
\title{Get the denominator of a 'ratioOfQsprays'}
\usage{
getDenominator(roq)
}
\arguments{
\item{roq}{a \code{ratioOfQsprays} object}
}
\value{
A \code{qspray} object.
}
\description{
Get the denominator of a \code{ratioOfQsprays} object,
  preserving the show options.
}
