% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{getNumerator}
\alias{getNumerator}
\title{Get the numerator of a 'ratioOfQsprays'}
\usage{
getNumerator(roq)
}
\arguments{
\item{roq}{a \code{ratioOfQsprays} object}
}
\value{
A \code{qspray} object.
}
\description{
Get the numerator of a \code{ratioOfQsprays} object,
  preserving the show options.
}
