% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{involvedVariables}
\alias{involvedVariables}
\alias{involvedVariables,ratioOfQsprays-method}
\title{Variables involved in a 'ratioOfQsprays'}
\usage{
\S4method{involvedVariables}{ratioOfQsprays}(x)
}
\arguments{
\item{x}{a \code{ratioOfQsprays} object}
}
\value{
A vector of integers. Each integer represents the index of a
  variable involved in \code{x}.
}
\description{
Variables involved in a \code{ratioOfQsprays} object.
}
\examples{
x <- qlone(1); z <- qlone(3)
rOQ <- 2*x/z + x/(x+z) + z^2/x
involvedVariables(rOQ) # should be c(1L, 3L)
}
\seealso{
\code{\link{numberOfVariables}}.
}
