% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{isConstant}
\alias{isConstant}
\alias{isConstant,ratioOfQsprays-method}
\title{Whether a 'ratioOfQsprays' is constant}
\usage{
\S4method{isConstant}{ratioOfQsprays}(x)
}
\arguments{
\item{x}{a \code{ratioOfQsprays} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a \code{ratioOfQsprays} object defines a constant
  fraction of polynomials.
}
