% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{numberOfVariables}
\alias{numberOfVariables}
\alias{numberOfVariables,ratioOfQsprays-method}
\title{Number of variables in a 'ratioOfQsprays'}
\usage{
\S4method{numberOfVariables}{ratioOfQsprays}(x)
}
\arguments{
\item{x}{a \code{ratioOfQsprays} object}
}
\value{
An integer.
}
\description{
Number of variables involved in a \code{ratioOfQsprays} object.
}
\note{
The number of variables in the \code{qspray} object \code{qlone(d)}
  is \code{d}, not \code{1}.
}
