% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioOfQsprays.R
\name{ratioOfQsprays-unary}
\alias{ratioOfQsprays-unary}
\alias{+,ratioOfQsprays,missing-method}
\alias{-,ratioOfQsprays,missing-method}
\title{Unary operators for 'ratioOfQsprays' objects}
\usage{
\S4method{+}{ratioOfQsprays,missing}(e1, e2)

\S4method{-}{ratioOfQsprays,missing}(e1, e2)
}
\arguments{
\item{e1}{object of class \code{ratioOfQsprays}}

\item{e2}{nothing}
}
\value{
A \code{ratioOfQsprays} object.
}
\description{
Unary operators for \code{ratioOfQsprays} objects.
}
