% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ratioOfQsprays_from_list}
\alias{ratioOfQsprays_from_list}
\title{(internal) Make a 'ratioOfQsprays' object from a list}
\usage{
ratioOfQsprays_from_list(x)
}
\arguments{
\item{x}{list returned by the Rcpp function
\code{returnRatioOfQsprays}}
}
\value{
A \code{ratioOfQsprays} object.
}
\description{
This function is for internal usage. It is exported because
  it is also used for internal usage in other packages.
}
