% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{rRatioOfQsprays}
\alias{rRatioOfQsprays}
\title{Random 'ratioOfQsprays'}
\usage{
rRatioOfQsprays()
}
\value{
A \code{ratioOfQsprays} object.
}
\description{
Generates a random \code{ratioOfQsprays} object.
}
