% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showRatioOfQsprays}
\alias{showRatioOfQsprays}
\title{Print a 'ratioOfQsprays' object}
\usage{
showRatioOfQsprays(showQspray, quotientBar = "  \%//\%  ")
}
\arguments{
\item{showQspray}{a function which prints a \code{qspray} object, which will
be applied to the numerator and the denominator}

\item{quotientBar}{a string representing the quotient bar between the
numerator and the denominator, including surrounding spaces,
e.g \code{" / "}}
}
\value{
A function which takes as argument a \code{ratioOfQsprays} object
  and which prints it.
}
\description{
Prints a \code{ratioOfQsprays} object given a function to print
  a \code{qspray} object
}
\seealso{
\code{\link{showRatioOfQspraysX1X2X3}},
  \code{\link{showRatioOfQspraysXYZ}},
  \code{\link{showRatioOfQspraysOption<-}}.
}
