% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showRatioOfQspraysUnivariate}
\alias{showRatioOfQspraysUnivariate}
\title{Title}
\usage{
showRatioOfQspraysUnivariate(var, quotientBar = "  \%//\%  ", ...)
}
\arguments{
\item{var}{xx}

\item{quotientBar}{xx}

\item{...}{xx}
}
\value{
xx
}
\description{
Title
}
