% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showRatioOfQspraysXYZ}
\alias{showRatioOfQspraysXYZ}
\title{Print a 'ratioOfQsprays'}
\usage{
showRatioOfQspraysXYZ(letters, quotientBar = "  \%//\%  ", ...)
}
\arguments{
\item{letters}{a vector of strings, usually some letters such as \code{"x"}
and \code{"y"}, to denote the variables}

\item{quotientBar}{a string representing the quotient bar between the
numerator and the denominator, including surrounding spaces,
e.g \code{" / "}}

\item{...}{arguments other than \code{quotientBar} passed to
\code{\link{showRatioOfQsprays}} (currently there's no such argument)}
}
\value{
A function which takes as argument a \code{ratioOfQsprays} object
  and which prints it.
}
\description{
Print a \code{ratioOfQsprays} object given some letters to
  denote the variables, by printing monomials like \code{"x^2yz"}.
}
\examples{
( roq <- rRatioOfQsprays() )
showRatioOfQspraysXYZ(c("X", "Y", "Z"), " / ")(roq)
}
