% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{substituteSomeRatioOfQsprays}
\alias{substituteSomeRatioOfQsprays}
\title{Substitute the variables of a 'ratioOfQsprays' with 
  some 'ratioOfQsprays'}
\usage{
substituteSomeRatioOfQsprays(roq, listOfRoqs)
}
\arguments{
\item{roq}{a \code{ratioOfQsprays} object}

\item{listOfRoqs}{the \code{ratioOfQsprays} to be substituted to the 
variables of \code{roq}; this must be a list of some objects coercible 
to some \code{ratioOfQsprays} objects (by an application of the function
\code{\link{as.ratioOfQsprays}}), and the length of this list must be 
equal to the number of variables of \code{roq}}
}
\value{
A \code{ratioOfQsprays} object.
}
\description{
Substitute the variables of a \code{ratioOfQsprays} fraction 
  of polynomials with some \code{ratioOfQsprays} objects, or some objects 
  coercible to some \code{ratioOfQsprays}.
}
\examples{
x <- qlone(1)
y <- qlone(2)
z <- qlone(3)
p <- x + y
q <- x - y
rOQ <- p / q
rOQ1 <- x / y
rOQ2 <- z / x
substituteSomeRatioOfQsprays(rOQ, list(rOQ1, rOQ2))
}
