% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{withAttributes}
\alias{withAttributes}
\title{Set show options to a 'ratioOfQsprays' object}
\usage{
withAttributes(roq, x = "x", quotientBar = "  \%//\%  ")
}
\arguments{
\item{roq}{a \code{ratioOfQsprays} object}

\item{x}{value for the \code{"x"} attribute}

\item{quotientBar}{value for the \code{"quotientBar"} attribute}
}
\value{
The input \code{ratioOfQsprays} object with new attributes.
}
\description{
Set some attributes to a \code{ratioOfQsprays} object
  to control the way it is displayed. See the note in the
  documentation of \code{\link[showRatioOfQspraysCanonical]} for details.
}
\examples{
( roq <- rRatioOfQsprays() )
withAttributes(roq, x = "a", quotientBar = " / ")
}
