% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{isUnivariate}
\alias{isUnivariate}
\alias{isUnivariate,ratioOfQsprays-method}
\title{Whether a 'ratioOfQsprays' is univariate}
\usage{
\S4method{isUnivariate}{ratioOfQsprays}(x)
}
\arguments{
\item{x}{a \code{ratioOfQsprays} object}
}
\value{
A Boolean value.
}
\description{
Checks whether a \code{ratioOfQsprays} object defines a
  univariate fraction of polynomials.
}
\note{
The \code{ratioOfQsprays} object \code{y / (1 + y)} where
  \code{y=qlone(2)} is not univariate, although it involves only one
  variable. The function returns \code{TRUE} when only \code{qlone(1)}
  is involved or when no variable is involved.
}
