% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{methods}
\name{numberOfVariables}
\alias{numberOfVariables}
\alias{numberOfVariables,ratioOfQsprays-method}
\title{Number of variables in a 'ratioOfQsprays'}
\usage{
\S4method{numberOfVariables}{ratioOfQsprays}(x)
}
\arguments{
\item{x}{a \code{ratioOfQsprays} object}
}
\value{
An integer.
}
\description{
Number of variables involved in a \code{ratioOfQsprays} object.
}
\note{
The number of variables in the \code{ratioOfQsprays} object
  \code{y / (1 + y)} where \code{y=qlone(2)} is \code{2}, not \code{1},
  although only one variable occurs. Rigorously speaking, the function
  returns the maximal integer \code{d} such that \code{qlone(d)} occurs in
  the 'ratioOfQsprays'.
}
\seealso{
\code{\link{involvedVariables}}
}
