% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{rRatioOfQsprays}
\alias{rRatioOfQsprays}
\title{Random 'ratioOfQsprays'}
\usage{
rRatioOfQsprays(allow.zero = TRUE)
}
\arguments{
\item{allow.zero}{Boolean, whether to allow to get a null
\code{ratioOfQsprays}}
}
\value{
A \code{ratioOfQsprays} object.
}
\description{
Generates a random \code{ratioOfQsprays} object.
}
