% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showRatioOfQspraysCanonical}
\alias{showRatioOfQspraysCanonical}
\title{Print a 'ratioOfQsprays'}
\usage{
showRatioOfQspraysCanonical(var, quotientBar = "  \%//\%  ", ...)
}
\arguments{
\item{var}{a string, usually a letter, to denote the unindexed variables}

\item{quotientBar}{a string representing the quotient bar between the
numerator and the denominator, including surrounding spaces,
e.g \code{" / "}}

\item{...}{arguments other than \code{quotientBar} passed to
\code{\link{showRatioOfQsprays}} (currently there's no such argument)}
}
\value{
A function which takes as argument a \code{ratioOfQsprays} object
  and which prints it.
}
\description{
Print a \code{ratioOfQsprays} object given a string to denote
  the unindexed variables.
}
\note{
The \code{show} method for \code{ratioOfQsprays} objects uses
  \code{showRatioOfQspraysCanonical("x", quotientBar = "  \%//\%  ")}
  by default. But this can be controlled as follows. If a
  \code{ratioOfQsprays} object has an attribute \code{"x"}, then the value
  of this attribute will replace \code{"x"} in the \code{show} output.
  It is also possible to control the \code{quotientBar} argument by
  assigning a \code{"quotientBar"} attribute to the \code{ratioOfQsprays}
  object to be printed.
}
\examples{
roq <- rRatioOfQsprays()
showRatioOfQspraysCanonical("X", " / ")(roq)
}
