% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showRatioOfQspraysOption<-}
\alias{showRatioOfQspraysOption<-}
\title{Set a show option to a 'ratioOfQsprays'}
\usage{
showRatioOfQspraysOption(x, which) <- value
}
\arguments{
\item{x}{a \code{ratioOfQsprays} object}

\item{which}{which option to set; this can be \code{"x"},
\code{"quotientBar"}, \code{"showQspray"}, or \code{"showRatioOfQsprays"}}

\item{value}{the value of the option to be set}
}
\value{
This returns the updated \code{ratioOfQsprays}.
}
\description{
Set a show option to a \code{ratioOfQsprays} object.
}
\examples{
set.seed(666)
( roq <- rRatioOfQsprays() )
showRatioOfQspraysOption(roq, "quotientBar") <- " / "
roq
showRatioOfQspraysOption(roq, "x") <- "a"
roq
showRatioOfQspraysOption(roq, "showQspray") <- showQsprayXYZ()
roq
}
\seealso{
\code{\link{showRatioOfQsprays}}.
}
